I don't understand anything anymore
Back to 20 secs with ideal quality. Like WTF.

+++ src/me/oak/imgcolor/Bag.java
@@ -9,5 +9,5 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor public class Bag<T> {
 
     public final T value;
-    public short amount;
+    public int amount;
 }
